+++ src/edwh/local_tasks/plugin.py
@@ -14,7 +14,7 @@ import requests
 import yarl
 from invoke import Context, task
 from packaging.version import parse as parse_package_version
-from termcolor import colored
+from termcolor import colored, cprint
 
 from ..meta import (
     Version,
@@ -76,11 +76,9 @@ class Plugin:
                     f"• {self.clean_name} ({self.installed_version} < {self.latest_version}) - {self.github_url}"
                 )
 
-            print(
-                colored(
-                    plugin_details,
-                    "yellow",
-                )
+            cprint(
+                plugin_details,
+                "yellow",
             )
         elif self.is_installed:
             if verbose:
@@ -90,11 +88,9 @@ class Plugin:
             else:
                 plugin_details = f"• {self.clean_name} - {self.github_url}"
 
-            print(
-                colored(
-                    plugin_details,
-                    "green",
-                )
+            cprint(
+                plugin_details,
+                "green",
             )
         else:
             if verbose:
@@ -104,11 +100,9 @@ class Plugin:
             else:
                 plugin_details = f"◦ {self.clean_name} - {self.github_url}"
 
-            print(
-                colored(
-                    plugin_details,
-                    "red",
-                )
+            cprint(
+                plugin_details,
+                "red",
             )
 
 
@@ -172,23 +166,19 @@ def list_plugins(c, verbose=False):
         print()
         s = "" if len(old_plugins) == 1 else "s"
         verb = "is" if len(old_plugins) == 1 else "are"
-        print(
-            colored(
-                f"{len(old_plugins)} plugin{s} {verb} out of date. "
-                f"Try `edwh self-update` to fix this "
-                f"or `edwh plugins --changelog` to see what's new.",
-                "yellow",
-            )
+        cprint(
+            f"{len(old_plugins)} plugin{s} {verb} out of date. "
+            f"Try `edwh self-update` to fix this "
+            f"or `edwh plugins --changelog` to see what's new.",
+            "yellow",
         )
 
     if not_all_installed:
         print()
-        print(
-            colored(
-                f"Tip: not all plugins are installed. "
-                f"For example, try `edwh plugin.add {not_all_installed}` or `edwh plugin.add all`",
-                "blue",
-            )
+        cprint(
+            f"Tip: not all plugins are installed. "
+            f"For example, try `edwh plugin.add {not_all_installed}` or `edwh plugin.add all`",
+            "blue",
         )
 
 
@@ -235,7 +225,7 @@ def remove_all(c):
 
 
 @task(aliases=("install",))
-def add(c, plugin_name: str):
+def add(c, plugin_names: str):
     """
     Install a new plugin
 
@@ -243,54 +233,62 @@ def add(c, plugin_name: str):
         c (Context): invoke ctx
         plugin_name: which plugin to add
     """
-    if plugin_name == "all":
+    if plugin_names == "all":
         return add_all(c)
 
     pip = _pip()
-    plugin_name = _require_affixes(plugin_name)
 
-    c.run(f"{pip} install {plugin_name}")
+    plugin_names = [_require_affixes(plugin_name.strip()) for plugin_name in plugin_names.split(",")]
+
+    c.run(f"{pip} install " + " ".join(plugin_names))
 
 
 @task(aliases=("upgrade",))
-def update(c, plugin_name: str, version: str = None):
+def update(c, plugin_names: str, version: str = None, verbose: bool = False):
     """
     Update a plugin (or 'all') to the latest version
 
     Args:
         c (Context): invoke ctx
-        plugin_name: the edwh plugin name (can be supplied without edwh- prefix or -plugin suffix)
+        plugin_names: the edwh plugin name (can be supplied without edwh- prefix or -plugin suffix)
         version: optional custom version string (e.g. 0.14.0b1 for a beta pre-release)
+        verbose: show which will would be installed for each plugin
     """
-    if plugin_name == "all":
+    if plugin_names == "all":
         from ..tasks import self_update
 
         return self_update(c)
 
     pip = _pip()
-    plugin_name = _require_affixes(plugin_name)
-    # if version is supplied, choose that. Otherwise use the latest
-    version = version or _get_latest_version_from_pypi(plugin_name)
 
-    c.run(f"{pip} install {plugin_name}=={version}")
+    plugins_with_version = []
+    for plugin_name in plugin_names.split(","):
+        plugin_name = _require_affixes(plugin_name.strip())
+        plugin_version = version or _get_latest_version_from_pypi(plugin_name)
+        plugins_with_version.append(f"{plugin_name}=={plugin_version}")
+
+    if verbose:
+        cprint(str(plugins_with_version), "blue")
+
+    c.run(f"{pip} install " + " ".join(plugins_with_version))
 
 
 @task(aliases=("uninstall",))
-def remove(c, plugin_name: str):
+def remove(c, plugin_names: str):
     """
     Remove a plugin (or 'all')
 
     Args:
         c (Context): invoke ctx
-        plugin_name: which plugin to remove
+        plugin_names: which plugin to remove
     """
-    if plugin_name == "all":
+    if plugin_names == "all":
         return remove_all(c)
     pip = _pip()
     # ensure the prefix and suffix exist, but not twice:
-    plugin_name = _require_affixes(plugin_name)
+    plugin_names = [_require_affixes(plugin_name.strip()) for plugin_name in plugin_names.split(",")]
 
-    c.run(f"{pip} uninstall --yes {plugin_name}")
+    c.run(f"{pip} uninstall --yes " + " ".join(plugin_names))
 
 
 GITHUB_RAW_URL = yarl.URL("https://raw.githubusercontent.com")
@@ -497,7 +495,7 @@ def display_changelogs(changelogs: dict[str, OrderedDict]):
     Final step of changelog(), uses the result of {package: sort_and_filter_changelog()}.
     """
     for package, history in changelogs.items():
-        print(colored(package, "red", attrs=["bold", "underline"]))
+        cprint(package, "red", attrs=["bold", "underline"])
         for version, changes in history.items():
             print("-", version)
             for change_type, change_descriptions in changes.items():